requirements.txt
fastapi 
uvicorn 
psycopg2-binary 
pydantic 
pydantic-geojson 
geojson
shapely